+++ lib/view/my_app.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+import 'my_app_home.dart';
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  static const String title = 'Explicit Animation';
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: title,
+      home: MyAppHome(),
+    );
+  }
+}
